Backend/requirements.txt
Django==5.2.7
djangorestframework==3.16.1
psycopg2==2.9.11
python-dotenv==1.1.1
tzdata==2025.2
asgiref==3.10.0
sqlparse==0.5.3
